refactor: persistence 관련 설정 파일 패키지 생성
