+++ src/main/java/com/ark/inflearnback/config/persistence/QuerydslConfig.java
@@ -1,4 +1,4 @@
-package com.ark.inflearnback.config;
+package com.ark.inflearnback.config.persistence;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import javax.persistence.EntityManager;
